+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## JCM (v2.0.0-beta.12) for (MTR 4.0.0-beta.16) has been released!
 
+**Changes**
+- JCM blocks now no longer have the potential break when water gets flowed towards the block
+
 **Fix:**
 - Minor GUI fixes